+++ src/Controller/Admin/UsersController.php
@@ -24,7 +24,7 @@ namespace MeCms\Controller\Admin;
 
 use Cake\Mailer\MailerAwareTrait;
 use MeCms\Controller\AppController;
-use MeCms\Utility\LoginLogger;
+use MeCms\Utility\LoginRecorder;
 
 /**
  * Users controller
@@ -100,7 +100,7 @@ class UsersController extends AppController
      * Views user
      * @param string $id User ID
      * @return void
-     * @uses MeCms\Utility\LoginLogger::get()
+     * @uses MeCms\Utility\LoginRecorder::read()
      */
     public function view($id = null)
     {
@@ -113,7 +113,7 @@ class UsersController extends AppController
         $this->set(compact('user'));
 
         if (config('users.login_log')) {
-            $loginLog = (new LoginLogger($id))->get();
+            $loginLog = (new LoginRecorder($id))->read();
 
             $this->set(compact('loginLog'));
         }
@@ -268,7 +268,7 @@ class UsersController extends AppController
     /**
      * Displays the login log
      * @return \Cake\Network\Response|null|void
-     * @uses MeCms\Utility\LoginLogger::get()
+     * @uses MeCms\Utility\LoginRecorder::read()
      */
     public function lastLogin()
     {
@@ -279,7 +279,7 @@ class UsersController extends AppController
             return $this->redirect(['_name' => 'admin']);
         }
 
-        $loginLog = (new LoginLogger($this->Auth->user('id')))->get();
+        $loginLog = (new LoginRecorder($this->Auth->user('id')))->read();
 
         $this->set(compact('loginLog'));
     }